moves chess_text.py from src to /
