+++ chess_text.py
